combined 4 bum_traffic tests in 2 tests & update traffic items
